// WADNR.Web/src/app/shared/generated/model/project-contacts-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectContactStepItem } from './project-contact-step-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectContactsStep { 
    ProjectID?: number;
    Contacts?: Array<ProjectContactStepItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectContactsStepForm { 
    ProjectID?: FormControl<number>;
    Contacts?: FormControl<Array<ProjectContactStepItem>>;
}

export class ProjectContactsStepFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Contacts = (value: FormControlState<Array<ProjectContactStepItem>> | Array<ProjectContactStepItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectContactStepItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}