// WADNR.Web/src/app/shared/generated/model/project-contacts-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectContactRequestItem } from './project-contact-request-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectContactsStepRequest { 
    Contacts?: Array<ProjectContactRequestItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectContactsStepRequestForm { 
    Contacts?: FormControl<Array<ProjectContactRequestItem>>;
}

export class ProjectContactsStepRequestFormControls { 
    public static Contacts = (value: FormControlState<Array<ProjectContactRequestItem>> | Array<ProjectContactRequestItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectContactRequestItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}