// WADNR.Web/src/app/shared/generated/model/project-contact-request-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectContactRequestItem { 
    ProjectPersonID?: number | null;
    PersonID?: number;
    ProjectPersonRelationshipTypeID?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectContactRequestItemForm { 
    ProjectPersonID?: FormControl<number>;
    PersonID?: FormControl<number>;
    ProjectPersonRelationshipTypeID?: FormControl<number>;
}

export class ProjectContactRequestItemFormControls { 
    public static ProjectPersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectPersonRelationshipTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}